✨ first naive audio processing

+++ src/load_data.py
@@ -6,7 +6,7 @@ def load_data(path='data/parsed.csv'):
     with open(path, 'r') as csvfile:
         reader = csv.reader(csvfile)
         for row in reader:
-            yield np.array(row)
+            yield np.array(row).astype(float)
 
 
 fetch_data_iter = load_data()
